done memory detection and bitmap but might access reserved memory

+++ headers/kernel/utils.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-void int_to_ascii(int, char*);
+void int_to_ascii(uint64_t, char*);
 void append(char s[], char n);
 void backspace(char s[]);
 int strcmp(char s1[], char s2[]);
